feat(verify): stderr summary, full sha256, --require-clean, --print-canonical-payload
- Move verify summary line from stdout to stderr to match snapshot's stream
  convention; stdout reserved for structured payload (--print-canonical-payload).
- Print full 64-char sha256 instead of 8-char prefix in both snapshot and
  verify summaries — auditors pin to the full digest.
- Add --require-clean: fails when manifest tool_version ends in +dirty.
  Engagement-grade auditors refuse +dirty evidence; flag closes the gate.
- Surface +dirty on stderr WARN (parallel to existing tool_version=dev WARN)
  even without --require-clean.
- Add --print-canonical-payload: re-emits the canonical (JCS) bytes from
  the embedded manifest to stdout, skipping integrity checks. Lets
  downstream signature workflows re-derive the payload without trusting
  the operator's --signing-payload sidecar.

+++ cmd/snapshot.go
@@ -135,7 +135,7 @@ Signing example:
 		}
 
 		abs, _ := filepath.Abs(out)
-		fmt.Fprintf(os.Stderr, "Wrote snapshot to %s (sha256=%s, scans=%d, format=%s)\n", abs, short(hash), len(ids), format)
+		fmt.Fprintf(os.Stderr, "Wrote snapshot to %s (sha256=%s, scans=%d, format=%s)\n", abs, hash, len(ids), format)
 		return nil
 	},
 }
